Refactor kudu itcase.
