+++ bitsail-cores/bitsail-core-flink-base/src/main/java/com/bytedance/bitsail/flink/core/delagate/converter/FlinkRowConvertSerializer.java
@@ -230,6 +230,12 @@ public class FlinkRowConvertSerializer implements RowSerializer<Row> {
       return new java.sql.Time(column.asDate().getTime());
     } else if (TypeInfos.SQL_TIMESTAMP_TYPE_INFO.getTypeClass() == typeInfoTypeClass) {
       return new java.sql.Timestamp(column.asDate().getTime());
+    } else if (TypeInfos.LOCAL_DATE_TYPE_INFO.getTypeClass() == typeInfoTypeClass) {
+      return column.asDate();
+    } else if (TypeInfos.LOCAL_TIME_TYPE_INFO.getTypeClass() == typeInfoTypeClass) {
+      return column.asDate();
+    } else if (TypeInfos.LOCAL_DATE_TIME_TYPE_INFO.getTypeClass() == typeInfoTypeClass) {
+      return column.asDate();
     } else if (BasicArrayTypeInfo.BINARY_TYPE_INFO.getTypeClass() == typeInfoTypeClass) {
       return column.asBytes();
     } else {
