--- bitsail-test/bitsail-connector-kudu-test/src/test/resources/kudu_to_print.json
@@ -1,50 +1,0 @@
-{
-  "job": {
-    "common": {
-      "job_id": 312,
-      "instance_id": 3124,
-      "job_name": "bitsail_kudu_to_print_test",
-      "user_name": "test"
-    },
-    "reader": {
-      "class": "com.bytedance.bitsail.connector.kudu.source.KuduSource",
-      "kudu_table_name": "kudu_test_table",
-      "kudu_master_address_list": "localhost:1234",
-      "read_mode": "READ_LATEST",
-      "split_config": "{\"name\": \"key\", \"lower_bound\": 0, \"upper_bound\": \"10000\", \"split_num\": 3}",
-      "columns": [
-        {
-          "name": "key",
-          "type": "int64"
-        },
-        {
-          "name": "field_boolean",
-          "type": "boolean"
-        },
-        {
-          "name": "field_int",
-          "type": "int"
-        },
-        {
-          "name": "field_double",
-          "type": "double"
-        },
-        {
-          "name": "field_date",
-          "type": "date"
-        },
-        {
-          "name": "field_string",
-          "type": "string"
-        },
-        {
-          "name": "field_binary",
-          "type": "binary"
-        }
-      ]
-    },
-    "writer": {
-      "class": "com.bytedance.bitsail.connector.legacy.print.sink.PrintSink"
-    }
-  }
-}
