+++ bitsail-test/bitsail-test-integration/bitsail-test-integration-kudu/src/test/resources/fake_to_kudu.json
@@ -1,0 +1,84 @@
+{
+  "job": {
+    "common": {
+      "job_id": -2413,
+      "job_name": "bitsail_test_integrationfake_to_kudu",
+      "instance_id": -20413,
+      "user_name": "user"
+    },
+    "reader": {
+      "class": "com.bytedance.bitsail.connector.fake.source.FakeSource",
+      "total_count": 5000,
+      "rate": 1000,
+      "null_percentage": 20,
+      "columns": [
+        {
+          "name": "key",
+          "type": "long",
+          "properties": "unique"
+        },
+        {
+          "name": "fake_boolean",
+          "type": "boolean"
+        },
+        {
+          "name": "fake_int",
+          "type": "int"
+        },
+        {
+          "name": "fake_double",
+          "type": "double"
+        },
+        {
+          "name": "fake_date",
+          "type": "date.date"
+        },
+        {
+          "name": "fake_string",
+          "type": "string"
+        },
+        {
+          "name": "fake_binary",
+          "type": "binary"
+        }
+      ]
+    },
+    "writer": {
+      "class": "com.bytedance.bitsail.connector.kudu.sink.KuduSink",
+      "kudu_worker_count": 2,
+      "columns": [
+        {
+          "name": "key",
+          "type": "long",
+          "properties": "unique"
+        },
+        {
+          "name": "field_boolean",
+          "type": "boolean"
+        },
+        {
+          "name": "field_int",
+          "type": "int"
+        },
+        {
+          "name": "field_double",
+          "type": "double"
+        },
+        {
+          "name": "field_date",
+          "type": "date"
+        },
+        {
+          "name": "field_string",
+          "type": "string",
+          "properties": "nullable"
+        },
+        {
+          "name": "field_binary",
+          "type": "binary",
+          "properties": "nullable"
+        }
+      ]
+    }
+  }
+}
