+++ bitsail-test/pom.xml
@@ -29,7 +29,6 @@
     <modules>
         <module>bitsail-test-end-to-end</module>
         <module>bitsail-connector-test</module>
-        <module>bitsail-connector-kudu-test</module>
         <module>bitsail-test-integration</module>
     </modules>
